+++ .gitignore
@@ -99,3 +99,7 @@ ENV/
 
 # mypy
 .mypy_cache/
+
+# sublime
+*.sublime-project
+*.sublime-workspace
