+++ tests/test_recurrent_fire.py
@@ -68,6 +68,31 @@ def test_recurrent_fire_single():
     assert actual == expected
 
 
+def test_recurrent_fire_nested():
+    # GIVEN
+    expected = [
+        fluent_initiate('water', [], 0),
+        action('ignite', ['sofa'], (1, 2)),
+        fluent_initiate('fire', [], 2),
+        action('eliminate', [], (2, 3)),
+        fluent_terminate('fire', [], 3),
+        fluent_terminate('water', [], 3),
+        action('ignite', ['bed'], (4, 5)),
+        fluent_initiate('fire', [], 5),
+        action('refill', [], (7, 8)),
+        fluent_initiate('water', [], 8),
+        action('eliminate', [], (8, 9)),
+        fluent_terminate('fire', [], 9),
+        fluent_terminate('water', [], 9),
+    ]
+
+    # WHEN
+    actual = run_pylps_test_program('recurrent_fire', 'nested')
+
+    # THEN
+    assert actual == expected
+
+
 def test_recurrent_fire_delay_success():
     # GIVEN
     expected = [
